# utils/swagger.json
{
    "info": {
        "description": "API backend para aplica\u00e7\u00e3o",
        "version": "1.0.0-oas3",
        "contact": {
            "email": "raynan.serafim@gmail.com"
        },
        "title": "Backend"
    },
    "servers": [
        {
            "description": "Server local",
            "url": "http://localhost:5000"
        }
    ],
...
    "paths": {
...
        "/usuario/add": {
            "put": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Adiciona um registro",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome para filtro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados necess\u00e1rios para a cria\u00e7\u00e3o do registro",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/view/{query_id}": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca o registro do banco se ele existir",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioModel"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/list": {
            "get": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Busca lista de registro existentes no banco",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome de usuario para filtro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UsuarioModel"
                                            }
                                        }
                                    },
                                    "required": [
                                        "count",
                                        "items"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/update/{query_id}": {
            "put": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Edita um registro",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "requestBody": {
                    "description": "Dados necess\u00e1rios para a edi\u00e7\u00e3o do registro",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/usuario/delete/{query_id}": {
            "delete": {
                "tags": [
                    "Rotas"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "summary": "Remove o registro do banco se ele existir",
                "parameters": [
                    {
                        "in": "path",
                        "name": "query_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "description": "Identifica\u00e7\u00e3o \u00fanica do registro"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "openapi": "3.0.2",
    "components": {
        "schemas": {
...
            "UsuarioModel": {
                "title": "UsuarioModel",
                "type": "object",
                "properties": {
                    "nome": {
                        "title": "Nome",
                        "maxLength": 255,
                        "type": "string"
                    },
                    "pis": {
                        "title": "Pis",
                        "maxLength": 50,
                        "type": "string"
                    },
                    "cpf": {
                        "title": "Cpf",
                        "maxLength": 50,
                        "type": "string"
                    },
                    "endereco_id": {
                        "title": "Endereco Id",
                        "type": "integer"
                    }
                }
            }
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}